refactor: split query subscription lifecycle
