+++ crates/jazz/src/node/query_eval.rs
@@ -200,146 +200,6 @@ fn public_result_member_occurrence_id(
         .transpose()
 }
 
-pub(crate) struct LocalMaintainedViewSubscription {
-    subscription: MultisinkSubscription,
-    _retained_prepared_plan: Option<SubscriptionPreparedPlan>,
-    maintained: MaintainedSubscriptionView,
-    terminal_schemas: MaintainedTerminalSchemas,
-    tables: BTreeMap<String, TableSchema>,
-    result_query: JazzQuery,
-    result_table: String,
-    result_schema_version: SchemaVersionId,
-    binding_view_key: BindingViewKey,
-    result_select: Option<Vec<String>>,
-    result_set: BTreeSet<ResultMemberEntry>,
-    authoritative_result_set: BTreeSet<ResultMemberEntry>,
-    authoritative_result_generation: u64,
-    result_payloads: BTreeMap<ResultMemberEntry, ResultMemberPayloadEntry>,
-    program_facts: BTreeSet<ProgramFactEntry>,
-    root_occurrence_ids: Vec<OutputOccurrenceId>,
-}
-
-impl LocalMaintainedViewSubscription {
-    pub(crate) fn terminal_root_layout(&self) -> Option<&crate::db::TerminalRootLayout> {
-        self.terminal_schemas.terminal_root_layout()
-    }
-}
-
-/// A plan retained solely to keep a maintained subscription graph alive.
-/// Its provenance is established by the compiler path that produced it, so a
-/// caller cannot relabel a ClientLocal plan as TrustedServing after the fact.
-pub(crate) struct SubscriptionPreparedPlan {
-    plan: PreparedQueryPlanHandle,
-    authorization_mode: QueryAuthorizationMode,
-}
-
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
-#[cfg(feature = "testing")]
-pub(crate) struct LocalMaintainedViewSubscriptionFootprint {
-    pub(crate) maintained: MaintainedSubscriptionViewFootprint,
-    pub(crate) terminal_schemas: MaintainedTerminalSchemasFootprint,
-    pub(crate) tables: usize,
-    pub(crate) result_set: usize,
-    pub(crate) result_payloads: usize,
-    pub(crate) program_facts: usize,
-    pub(crate) control_state_bytes: usize,
-    pub(crate) total_heap_bytes: usize,
-}
-
-impl LocalMaintainedViewSubscription {
-    pub(crate) fn subscription_id(&self) -> groove::ivm::SubscriptionId {
-        self.subscription.id()
-    }
-
-    pub(crate) fn root_occurrence_ids(&self) -> &[OutputOccurrenceId] {
-        &self.root_occurrence_ids
-    }
-
-    #[cfg(test)]
-    pub(crate) fn retained_plan_authorization_mode(&self) -> Option<QueryAuthorizationMode> {
-        self._retained_prepared_plan
-            .as_ref()
-            .map(|plan| plan.authorization_mode)
-    }
-
-    #[cfg(feature = "testing")]
-    pub(crate) fn footprint(&self) -> LocalMaintainedViewSubscriptionFootprint {
-        let maintained = self.maintained.footprint();
-        let terminal_schemas = self.terminal_schemas.footprint();
-        let tables_bytes = self
-            .tables
-            .iter()
-            .map(|(name, schema)| name.len() + std::mem::size_of_val(schema))
-            .sum::<usize>()
-            + self.tables.len() * 96;
-        let result_set_bytes = self
-            .result_set
-            .iter()
-            .map(|member| {
-                postcard::to_allocvec(member)
-                    .map(|bytes| bytes.len())
-                    .unwrap_or(0)
-            })
-            .sum::<usize>()
-            + self.result_set.len() * 64;
-        let authoritative_result_set_bytes = self
-            .authoritative_result_set
-            .iter()
-            .map(|member| {
-                postcard::to_allocvec(member)
-                    .map(|bytes| bytes.len())
-                    .unwrap_or(0)
-            })
-            .sum::<usize>()
-            + self.authoritative_result_set.len() * 64;
-        let result_payloads_bytes = self
-            .result_payloads
-            .iter()
-            .map(|(member, payload)| {
-                postcard::to_allocvec(member)
-                    .map(|bytes| bytes.len())
-                    .unwrap_or(0)
-                    + postcard::to_allocvec(payload)
-                        .map(|bytes| bytes.len())
-                        .unwrap_or(0)
-            })
-            .sum::<usize>()
-            + self.result_payloads.len() * 96;
-        let program_facts_bytes = self
-            .program_facts
-            .iter()
-            .map(|fact| {
-                postcard::to_allocvec(fact)
-                    .map(|bytes| bytes.len())
-                    .unwrap_or(0)
-            })
-            .sum::<usize>()
-            + self.program_facts.len() * 64;
-        let control_state_bytes = terminal_schemas.terminal_schemas_bytes
-            + tables_bytes
-            + self.result_table.len()
-            + self
-                .result_select
-                .as_ref()
-                .map(|columns| columns.iter().map(String::len).sum::<usize>())
-                .unwrap_or_default()
-            + result_set_bytes
-            + authoritative_result_set_bytes
-            + result_payloads_bytes
-            + program_facts_bytes;
-        LocalMaintainedViewSubscriptionFootprint {
-            maintained,
-            terminal_schemas,
-            tables: self.tables.len(),
-            result_set: self.result_set.len(),
-            result_payloads: self.result_payloads.len(),
-            program_facts: self.program_facts.len(),
-            control_state_bytes,
-            total_heap_bytes: maintained.total_heap_bytes + control_state_bytes,
-        }
-    }
-}
-
 pub(crate) fn take_required_sink_deltas(
     mut deltas: MultisinkDeltas,
     sink: &str,
@@ -353,16 +213,6 @@ mod lowering;
 
 pub(crate) use lowering::PolicyAuthorizationGraph;
 use lowering::*;
-
-pub(crate) struct LocalMaintainedViewSubscriptionUpdate {
-    pub(crate) authoritative_membership_changed: bool,
-    pub(crate) added: Vec<(OutputOccurrenceId, CurrentRow)>,
-    pub(crate) removed: Vec<OutputOccurrenceId>,
-    pub(crate) added_edges: Vec<(RelationEdge, Option<CurrentRow>)>,
-    pub(crate) removed_edges: Vec<RelationEdge>,
-    pub(crate) terminal_operations: Vec<groove::ivm::TerminalOperation>,
-    pub(crate) terminal_layout: Option<crate::db::TerminalRootLayout>,
-}
 
 enum CurrentQueryProgramOutput {
     AppRows,
@@ -379,6 +229,17 @@ use source_resolution::*;
 mod normalization;
 
 use normalization::*;
+
+mod subscriptions;
+
+mod maintained_views;
+
+#[cfg(feature = "testing")]
+pub(crate) use maintained_views::LocalMaintainedViewSubscriptionFootprint;
+use maintained_views::SubscriptionPreparedPlan;
+pub(crate) use maintained_views::{
+    LocalMaintainedViewSubscription, LocalMaintainedViewSubscriptionUpdate,
+};
 
 impl<S> NodeState<S>
 where
@@ -522,644 +383,6 @@ where
             program_binding.values = values;
         }
         Ok(program_binding)
-    }
-
-    pub(super) fn register_shape(&mut self, shape_id: ShapeId, ast: ShapeAst) -> Result<(), Error> {
-        if ast.version != ShapeAst::VERSION {
-            return Err(Error::InvalidStoredValue("unsupported query AST version"));
-        }
-        let schema = if ast.schema_version == self.catalogue.current_schema_version_id {
-            &self.catalogue.schema
-        } else {
-            let Some(schema) = self.catalogue.catalogue_schemas.get(&ast.schema_version) else {
-                self.sync_metrics.parked_catalogue_shapes += 1;
-                self.parking
-                    .parked_shape_registrations
-                    .insert(shape_id, ast);
-                return Ok(());
-            };
-            &schema.schema
-        };
-        let shape = match &ast.body {
-            ShapeBody::Query(query) => {
-                query.validate_with_schema_version(schema, ast.schema_version)?
-            }
-            ShapeBody::Relation(relation) => relation_query_to_query(relation)?
-                .validate_with_schema_version(schema, ast.schema_version)?,
-        };
-        if shape.shape_id() != shape_id {
-            return Err(Error::InvalidStoredValue("shape id does not match AST"));
-        }
-        self.query.registered_shapes.insert(shape_id, shape);
-        self.drain_parked_binding_deltas_for_shape(shape_id)?;
-        Ok(())
-    }
-
-    pub(crate) fn validate_shape_ast_for_registration(
-        &self,
-        shape_id: ShapeId,
-        ast: &ShapeAst,
-    ) -> Result<Option<ValidatedQuery>, Error> {
-        if ast.version != ShapeAst::VERSION {
-            return Err(Error::InvalidStoredValue("unsupported query AST version"));
-        }
-        let schema = if ast.schema_version == self.catalogue.current_schema_version_id {
-            &self.catalogue.schema
-        } else {
-            let Some(schema) = self.catalogue.catalogue_schemas.get(&ast.schema_version) else {
-                return Ok(None);
-            };
-            &schema.schema
-        };
-        let shape = match &ast.body {
-            ShapeBody::Query(query) => {
-                query.validate_with_schema_version(schema, ast.schema_version)?
-            }
-            ShapeBody::Relation(relation) => relation_query_to_query(relation)?
-                .validate_with_schema_version(schema, ast.schema_version)?,
-        };
-        if shape.shape_id() != shape_id {
-            return Err(Error::InvalidStoredValue("shape id does not match AST"));
-        }
-        Ok(Some(shape))
-    }
-
-    pub(super) fn drain_parked_shape_registrations(&mut self) -> Result<(), Error> {
-        let ready = self
-            .parking
-            .parked_shape_registrations
-            .iter()
-            .filter_map(|(shape_id, ast)| {
-                self.catalogue
-                    .catalogue_schemas
-                    .contains_key(&ast.schema_version)
-                    .then_some((*shape_id, ast.clone()))
-            })
-            .collect::<Vec<_>>();
-        for (shape_id, ast) in ready {
-            self.parking.parked_shape_registrations.remove(&shape_id);
-            self.sync_metrics.parked_catalogue_shapes_resolved += 1;
-            self.register_shape(shape_id, ast)?;
-        }
-        Ok(())
-    }
-
-    pub(super) fn apply_subscribe(&mut self, subscribe: Subscribe) -> Result<(), Error> {
-        let Some(shape) = self
-            .query
-            .registered_shapes
-            .get(&subscribe.shape_id)
-            .cloned()
-        else {
-            self.parking
-                .parked_binding_deltas
-                .entry(subscribe.shape_id)
-                .or_default()
-                .push(subscribe);
-            return Ok(());
-        };
-        self.apply_known_shape_subscribe(&shape, subscribe)
-    }
-
-    pub(crate) fn register_query_subscription_for_peer(
-        &mut self,
-        shape_id: ShapeId,
-        ast: ShapeAst,
-        subscribe: Subscribe,
-    ) -> Result<(), Error> {
-        self.register_shape(shape_id, ast)?;
-        self.apply_subscribe(subscribe)
-    }
-
-    fn drain_parked_binding_deltas_for_shape(&mut self, shape_id: ShapeId) -> Result<(), Error> {
-        let Some(deltas) = self.parking.parked_binding_deltas.remove(&shape_id) else {
-            return Ok(());
-        };
-        let Some(shape) = self.query.registered_shapes.get(&shape_id).cloned() else {
-            self.parking.parked_binding_deltas.insert(shape_id, deltas);
-            return Ok(());
-        };
-        for subscribe in deltas {
-            self.apply_known_shape_subscribe(&shape, subscribe)?;
-        }
-        Ok(())
-    }
-
-    fn apply_known_shape_subscribe(
-        &mut self,
-        shape: &ValidatedQuery,
-        subscribe: Subscribe,
-    ) -> Result<(), Error> {
-        if subscribe.values.len() != shape.params().len() {
-            return Err(Error::InvalidStoredValue("binding arity mismatch"));
-        }
-        let value_map = shape
-            .params()
-            .keys()
-            .cloned()
-            .zip(subscribe.values.iter().cloned())
-            .collect::<BTreeMap<_, _>>();
-        let binding = shape.bind(value_map)?;
-        let binding_view_key = BindingViewKey {
-            shape_id: subscribe.shape_id,
-            binding_id: binding.binding_id(),
-            read_view: subscribe.subscription.read_view,
-        };
-        if subscribe.known_state.is_some() {
-            self.query
-                .known_state_declared_binding_views
-                .insert(binding_view_key);
-        } else {
-            self.query
-                .known_state_declared_binding_views
-                .remove(&binding_view_key);
-        }
-        self.query
-            .registered_bindings
-            .entry(subscribe.shape_id)
-            .or_default()
-            .insert(
-                subscribe.subscription.binding_id,
-                RegisteredBinding {
-                    values: subscribe.values,
-                    read_view: subscribe.subscription.read_view,
-                    binding_view_key,
-                },
-            );
-        Ok(())
-    }
-
-    pub(crate) fn apply_unsubscribe(&mut self, subscription: SubscriptionKey) {
-        let binding_view_key = self.binding_view_key_for_subscription(subscription).ok();
-        if let Some(bindings) = self
-            .query
-            .registered_bindings
-            .get_mut(&subscription.shape_id)
-        {
-            bindings.remove(&subscription.binding_id);
-        }
-        if let Some(binding_view_key) = binding_view_key
-            && !self.registered_binding_resolves_to_binding_view_key(binding_view_key)
-        {
-            self.clear_settled_result_view(binding_view_key);
-            self.query.settled_program_facts.remove(&binding_view_key);
-            self.query
-                .known_state_declared_binding_views
-                .remove(&binding_view_key);
-            self.query
-                .initial_hydration_binding_views
-                .remove(&binding_view_key);
-            self.query
-                .pending_opening_binding_views
-                .remove(&binding_view_key);
-        }
-    }
-
-    fn registered_binding_resolves_to_binding_view_key(
-        &self,
-        binding_view_key: BindingViewKey,
-    ) -> bool {
-        let Some(bindings) = self
-            .query
-            .registered_bindings
-            .get(&binding_view_key.shape_id)
-        else {
-            return false;
-        };
-        bindings.values().any(|registered| {
-            if registered.read_view != binding_view_key.read_view {
-                return false;
-            }
-            registered.binding_view_key == binding_view_key
-        })
-    }
-
-    pub(crate) fn has_settled_result_set(&self, binding_view_key: BindingViewKey) -> bool {
-        self.query
-            .settled_result_sets
-            .contains_key(&binding_view_key)
-    }
-
-    pub(crate) fn applied_view_update_generation(&self, binding_view_key: BindingViewKey) -> u64 {
-        self.query
-            .applied_view_update_generations
-            .get(&binding_view_key)
-            .copied()
-            .unwrap_or_default()
-    }
-
-    #[cfg(test)]
-    pub(crate) fn reset_subscription_snapshot_for_link_call_count(&mut self) {
-        SUBSCRIPTION_SNAPSHOT_FOR_LINK_CALLS.with(|calls| calls.set(0));
-    }
-
-    #[cfg(test)]
-    pub(crate) fn subscription_snapshot_for_link_call_count(&self) -> usize {
-        SUBSCRIPTION_SNAPSHOT_FOR_LINK_CALLS.with(std::cell::Cell::get)
-    }
-
-    #[cfg(test)]
-    pub(crate) fn inject_pending_authoritative_reset_for_test(
-        &mut self,
-        binding_view_key: BindingViewKey,
-        members: impl IntoIterator<Item = ResultMemberEntry>,
-        settled_through: GlobalSeq,
-    ) {
-        self.clear_settled_result_view(binding_view_key);
-        for member in members {
-            self.insert_settled_result_member_indexed(binding_view_key, member);
-        }
-        self.query
-            .settled_through_by_binding_view
-            .insert(binding_view_key, settled_through);
-        self.query
-            .pending_authoritative_reset_binding_views
-            .insert(binding_view_key);
-    }
-
-    pub(crate) fn take_pending_authoritative_reset_binding_views(
-        &mut self,
-    ) -> BTreeSet<BindingViewKey> {
-        std::mem::take(&mut self.query.pending_authoritative_reset_binding_views)
-    }
-
-    pub(crate) fn take_pending_terminal_operations(
-        &mut self,
-        binding_view_key: BindingViewKey,
-    ) -> Vec<groove::ivm::TerminalOperation> {
-        self.query
-            .pending_terminal_operations_by_binding_view
-            .remove(&binding_view_key)
-            .unwrap_or_default()
-    }
-
-    pub(crate) fn defer_authoritative_reset_for_binding_view(
-        &mut self,
-        binding_view_key: BindingViewKey,
-    ) {
-        self.query
-            .pending_authoritative_reset_binding_views
-            .insert(binding_view_key);
-    }
-
-    #[cfg(test)]
-    pub(crate) fn has_pending_authoritative_reset_for_test(
-        &self,
-        binding_view_key: BindingViewKey,
-    ) -> bool {
-        self.query
-            .pending_authoritative_reset_binding_views
-            .contains(&binding_view_key)
-    }
-
-    pub(crate) fn publication_deferred_for_binding_view(
-        &self,
-        binding_view_key: BindingViewKey,
-    ) -> bool {
-        self.query
-            .deferred_publication_binding_views
-            .contains(&binding_view_key)
-    }
-
-    pub(crate) fn opening_pending_for_binding_view(
-        &self,
-        binding_view_key: BindingViewKey,
-    ) -> bool {
-        self.query
-            .pending_opening_binding_views
-            .contains(&binding_view_key)
-    }
-
-    pub(crate) fn settled_result_transitions_for_subscription(
-        &self,
-        subscription: SubscriptionKey,
-        previous_member_result_set: &BTreeSet<ResultMemberEntry>,
-        previous_program_fact_set: &BTreeSet<ProgramFactEntry>,
-        result_table_filter: Option<&str>,
-        output_tables: &BTreeMap<String, TableSchema>,
-    ) -> Result<Option<super::maintained_subscription_view::ResultTransitions>, Error> {
-        let binding_view_key = self.binding_view_key_for_subscription(subscription)?;
-        // Settled binding views are shared by canonical query binding, while a
-        // table read policy is identity-scoped. Never relay a synthetic
-        // aggregate from that shared cache across an identity boundary; the
-        // per-peer maintained program remains the authority for policy-shaped
-        // aggregate output.
-        let shared_view_has_read_policy = self
-            .query
-            .registered_shapes
-            .get(&subscription.shape_id)
-            .and_then(|shape| self.table(shape.query().table.as_str()).ok())
-            .is_some_and(|table| table.read_policy.is_some());
-        let Some(settled_members) = self.query.settled_result_sets.get(&binding_view_key) else {
-            return Ok(None);
-        };
-        let settled_facts = self
-            .query
-            .settled_program_facts
-            .get(&binding_view_key)
-            .cloned()
-            .unwrap_or_default();
-        let member_is_visible = |member: &ResultMemberEntry| {
-            let Some(table_name) = member.table_name() else {
-                return false;
-            };
-            result_table_filter.is_none_or(|table| table_name == table)
-                && (output_tables.contains_key(table_name)
-                    || (matches!(member, ResultMemberEntry::Synthetic { .. })
-                        && !shared_view_has_read_policy))
-        };
-        let current = settled_members
-            .iter()
-            .filter(|member| member_is_visible(member))
-            .cloned()
-            .collect::<BTreeSet<_>>();
-        let previous = previous_member_result_set
-            .iter()
-            .filter(|member| member_is_visible(member))
-            .cloned()
-            .collect::<BTreeSet<_>>();
-        let fact_is_visible = |fact: &ProgramFactEntry| match fact {
-            ProgramFactEntry::ResultPayload(payload) => member_is_visible(&payload.member),
-            _ => true,
-        };
-        let current_facts = settled_facts
-            .into_iter()
-            .filter(fact_is_visible)
-            .collect::<BTreeSet<_>>();
-        let previous_facts = previous_program_fact_set
-            .iter()
-            .filter(|fact| fact_is_visible(fact))
-            .cloned()
-            .collect::<BTreeSet<_>>();
-        let program_fact_adds = current_facts
-            .difference(&previous_facts)
-            .cloned()
-            .collect::<Vec<_>>();
-        let program_fact_removes = previous_facts
-            .difference(&current_facts)
-            .cloned()
-            .collect::<Vec<_>>();
-        // A synthetic aggregate member is meaningful only together with its
-        // payload fact. In particular, an empty aggregate has a member and a
-        // payload whose aggregate field is `Nullable(None)`; it is not a
-        // member with a missing payload. Carry both representations through
-        // the settled-view handoff so facade materialization can retain that
-        // distinction.
-        let result_payload_adds = program_fact_adds
-            .iter()
-            .filter_map(|fact| match fact {
-                ProgramFactEntry::ResultPayload(payload) => {
-                    Some((payload.member.clone(), payload.clone()))
-                }
-                _ => None,
-            })
-            .collect();
-        let result_payload_removes = program_fact_removes
-            .iter()
-            .filter_map(|fact| match fact {
-                ProgramFactEntry::ResultPayload(payload) => Some(payload.member.clone()),
-                _ => None,
-            })
-            .collect();
-        Ok(Some(
-            super::maintained_subscription_view::ResultTransitions {
-                authoritative_membership_changed: false,
-                authoritative_member_adds: BTreeSet::new(),
-                adds: current.difference(&previous).cloned().collect(),
-                removes: previous.difference(&current).cloned().collect(),
-                result_payload_adds,
-                result_payload_removes,
-                program_fact_adds,
-                program_fact_removes,
-                structured_app_row_changes: BTreeSet::new(),
-                allow_storage_witness_fallback: true,
-                observed_result_delta_batches: 0,
-                requires_authoritative_membership_reconcile: false,
-                terminal_operations: Vec::new(),
-            },
-        ))
-    }
-
-    pub(crate) fn authoritative_reset_snapshot_for_binding_view(
-        &mut self,
-        shape: &ValidatedQuery,
-        binding_view_key: BindingViewKey,
-    ) -> Result<Option<RelationSnapshot>, Error> {
-        let Some(result_members) = self
-            .query
-            .settled_result_sets
-            .get(&binding_view_key)
-            .cloned()
-        else {
-            return Ok(None);
-        };
-        let program_facts = self
-            .query
-            .settled_program_facts
-            .get(&binding_view_key)
-            .cloned()
-            .unwrap_or_default();
-        let result_payloads = program_facts
-            .iter()
-            .filter_map(|fact| match fact {
-                ProgramFactEntry::ResultPayload(payload) => {
-                    Some((payload.member.clone(), payload.clone()))
-                }
-                _ => None,
-            })
-            .collect::<BTreeMap<_, _>>();
-
-        let result_table = shape.query().table.as_str();
-        let mut rows = Vec::new();
-        let mut row_keys = BTreeSet::new();
-        for member in result_members.iter().filter(|member| {
-            is_public_result_member(member, result_table, shape.query().aggregate.is_some())
-        }) {
-            let Some(row) = self.materialize_authoritative_reset_member(
-                shape.query(),
-                member,
-                &result_payloads,
-            )?
-            else {
-                continue;
-            };
-            row_keys.insert((row.table().to_owned(), row.row_uuid()));
-            rows.push(row);
-        }
-        // Result-member ordering is for identity and deduplication, not public
-        // query rank. Membership/windowing is already lowered; only restore the
-        // selected roots to their advertised order before sending a reset.
-        self.apply_query_order_in_schema(shape.query(), shape.schema_version(), &mut rows)?;
-        if shape.query().flat_join.is_none() {
-            self.apply_projection_in_schema(shape.query(), shape.schema_version(), &mut rows)?;
-        }
-        let root_count = rows.len();
-        let mut edges = Vec::new();
-        for fact in program_facts {
-            let ProgramFactEntry::RelationEdge(edge) = fact else {
-                continue;
-            };
-            // Program facts retain canonical authored identity.  The public
-            // relation snapshot, including its removal index, is keyed in the
-            // subscription read schema; project the edge identity alongside
-            // the row it references rather than mixing canonical `users`
-            // with a materialized `people` row.
-            let read_edge =
-                self.project_relation_edge_through_read_schema(&edge, shape.schema_version())?;
-            if row_keys.insert((read_edge.target_table.clone(), read_edge.target_row))
-                && let Some(version) = &edge.target_version
-                && let Some(row) = self.materialize_authoritative_reset_relation_edge_target(
-                    shape.schema_version(),
-                    edge.target_table.as_str(),
-                    edge.target_row,
-                    version,
-                )?
-            {
-                rows.push(row);
-            }
-            edges.push(read_edge);
-        }
-        Ok(Some(RelationSnapshot {
-            root_count,
-            rows,
-            edges,
-        }))
-    }
-
-    pub(crate) fn settled_through_for_binding_view(
-        &self,
-        binding_view_key: BindingViewKey,
-    ) -> Option<GlobalSeq> {
-        self.query
-            .settled_through_by_binding_view
-            .get(&binding_view_key)
-            .copied()
-    }
-
-    pub(crate) fn known_state_declaration_for_subscription(
-        &mut self,
-        shape: &ValidatedQuery,
-        binding: &Binding,
-        subscription: SubscriptionKey,
-        values: &[Value],
-        identity: AuthorId,
-    ) -> Result<Option<KnownStateDeclaration>, Error> {
-        let binding_view_key = BindingViewKey {
-            shape_id: shape.shape_id(),
-            binding_id: binding.binding_id(),
-            read_view: subscription.read_view,
-        };
-        if !self.has_settled_result_set(binding_view_key) {
-            let _ = self.load_known_state_fact(binding_view_key)?;
-            // Slow exact declarations are still known-state declarations: they
-            // must describe a binding view the server has previously settled
-            // for this client. A purely local first subscription could include
-            // rows the serving peer has not observed yet; truncating that to an
-            // exact set would silently overclaim and can make stale rehydrate
-            // responses suppress local live state.
-            return Ok(None);
-        }
-        if let Some(position) = self.settled_through_for_binding_view(binding_view_key) {
-            let authorization_progress = self
-                .query
-                .authorization_progress_by_binding_view
-                .get(&binding_view_key)
-                .copied();
-            return Ok(Some(match authorization_progress {
-                Some(authorization_progress) => {
-                    KnownStateDeclaration::FastWithAuthorizationProgress {
-                        completeness: KnownStateCompleteness::FastCurrentMembership,
-                        position,
-                        authorization_progress,
-                    }
-                }
-                None => KnownStateDeclaration::Fast {
-                    completeness: KnownStateCompleteness::FastCurrentMembership,
-                    position,
-                },
-            }));
-        }
-        if let Some(position) = self.load_known_state_fact(binding_view_key)? {
-            return Ok(Some(KnownStateDeclaration::Fast {
-                completeness: KnownStateCompleteness::FastCurrentMembership,
-                position,
-            }));
-        }
-        let mut refs = Vec::new();
-        for row in self.query_rows_for_link(shape, binding, DurabilityTier::Local, identity)? {
-            let Some(tx_id) = self.current_row_tx_id(&row) else {
-                continue;
-            };
-            refs.push(RowVersionRef::new(
-                row.table().to_owned(),
-                row.row_uuid(),
-                tx_id,
-            ));
-        }
-        refs.sort();
-        refs.dedup();
-        if refs.is_empty() {
-            return Ok(None);
-        }
-        Ok(exact_known_state_declaration_if_within_limits(
-            shape.shape_id(),
-            subscription,
-            values,
-            refs,
-        ))
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn subscription_is_known_state_declared(
-        &self,
-        subscription: SubscriptionKey,
-    ) -> Result<bool, Error> {
-        let binding_view_key = match self.binding_view_key_for_subscription(subscription) {
-            Ok(binding_view_key) => binding_view_key,
-            Err(Error::InvalidStoredValue(
-                "subscription referenced unregistered shape"
-                | "subscription referenced unregistered binding",
-            )) => return Ok(false),
-            Err(error) => return Err(error),
-        };
-        Ok(self
-            .query
-            .known_state_declared_binding_views
-            .contains(&binding_view_key))
-    }
-
-    pub(crate) fn binding_view_key_for_subscription(
-        &self,
-        subscription: SubscriptionKey,
-    ) -> Result<BindingViewKey, Error> {
-        if let Some(registered) = self
-            .query
-            .registered_bindings
-            .get(&subscription.shape_id)
-            .and_then(|bindings| bindings.get(&subscription.binding_id))
-        {
-            return Ok(registered.binding_view_key);
-        }
-        if let Some(binding_view_key) = self.canonical_whole_table_binding_view_key(subscription)? {
-            return Ok(binding_view_key);
-        }
-        Err(Error::InvalidStoredValue(
-            "subscription referenced unregistered binding",
-        ))
-    }
-
-    fn canonical_whole_table_binding_view_key(
-        &self,
-        subscription: SubscriptionKey,
-    ) -> Result<Option<BindingViewKey>, Error> {
-        for table in &self.catalogue.schema.tables {
-            if self.whole_table_subscription_key(&table.name)? == subscription {
-                return Ok(Some(BindingViewKey::from_canonical_subscription_key(
-                    subscription,
-                )));
-            }
-        }
-        Ok(None)
     }
 
     fn compile_current_query_program(
@@ -3015,684 +2238,6 @@ where
                 &[Value::U64(table_id.0), Value::U64(position.0 + 1)],
             )?)
         }
-    }
-
-    pub(crate) fn open_maintained_view_subscription_in_authorization_mode(
-        &mut self,
-        shape: &ValidatedQuery,
-        binding: &Binding,
-        identity: AuthorId,
-        tier: DurabilityTier,
-        read_view: &ReadViewSpec,
-        retained_prepared_plan: Option<SubscriptionPreparedPlan>,
-        authorization_mode: QueryAuthorizationMode,
-    ) -> Result<(LocalMaintainedViewSubscription, RelationSnapshot), Error> {
-        if let Some(retained) = retained_prepared_plan.as_ref() {
-            if retained.authorization_mode != authorization_mode {
-                return Err(Error::InvalidStoredValue(
-                    "maintained subscription retained a plan from another authorization mode",
-                ));
-            }
-            debug_assert!(std::sync::Arc::strong_count(&retained.plan) > 0);
-        }
-        let settled_binding_view = (authorization_mode == QueryAuthorizationMode::ClientLocal)
-            .then(|| {
-                self.client_settled_binding_view_key_for_query(shape, binding, tier, read_view)
-            })
-            .flatten();
-        let (subscription, maintained, terminal_schemas, transitions, tables) = self
-            .open_seeded_maintained_subscription_view_in_authorization_mode(
-                shape,
-                binding,
-                identity,
-                tier,
-                read_view,
-                authorization_mode,
-                settled_binding_view,
-                PreparedClaimBindingMode::Strict,
-            )?;
-        let mut local = LocalMaintainedViewSubscription {
-            subscription,
-            _retained_prepared_plan: retained_prepared_plan,
-            maintained,
-            terminal_schemas,
-            tables,
-            result_query: shape.query().clone(),
-            result_table: shape.query().table.clone(),
-            result_schema_version: shape.schema_version(),
-            binding_view_key: settled_binding_view.unwrap_or_else(|| {
-                BindingViewKey::new(
-                    shape.shape_id(),
-                    binding.binding_id(),
-                    RegisterShapeOptions {
-                        tier,
-                        read_view: read_view.clone(),
-                        ..RegisterShapeOptions::default()
-                    }
-                    .read_view_key(),
-                )
-            }),
-            result_select: shape.query().select.clone(),
-            result_set: BTreeSet::new(),
-            authoritative_result_set: BTreeSet::new(),
-            authoritative_result_generation: 0,
-            result_payloads: BTreeMap::new(),
-            program_facts: BTreeSet::new(),
-            root_occurrence_ids: Vec::new(),
-        };
-        let _initial_delta =
-            self.apply_local_maintained_view_transitions(&mut local, transitions)?;
-        let initial =
-            self.materialize_local_maintained_relation_snapshot_with_occurrences(&local)?;
-        local.root_occurrence_ids = initial.root_occurrence_ids;
-        Ok((local, initial.snapshot))
-    }
-
-    pub(crate) fn drain_local_maintained_view_subscription(
-        &mut self,
-        local: &mut LocalMaintainedViewSubscription,
-        authoritative_binding_view: Option<BindingViewKey>,
-    ) -> Result<Option<LocalMaintainedViewSubscriptionUpdate>, Error> {
-        let Some(transitions) = self.drain_local_maintained_view_subscription_transitions(
-            local,
-            authoritative_binding_view,
-        )?
-        else {
-            return Ok(None);
-        };
-        let update = self.apply_local_maintained_view_transitions(local, transitions)?;
-        Ok(Some(update))
-    }
-
-    pub(crate) fn drain_local_maintained_view_subscription_state(
-        &mut self,
-        local: &mut LocalMaintainedViewSubscription,
-        authoritative_binding_view: Option<BindingViewKey>,
-    ) -> Result<bool, Error> {
-        let Some(transitions) = self.drain_local_maintained_view_subscription_transitions(
-            local,
-            authoritative_binding_view,
-        )?
-        else {
-            return Ok(false);
-        };
-        let _ = self.apply_local_maintained_view_transitions_inner(local, transitions, false)?;
-        Ok(true)
-    }
-
-    pub(crate) fn reset_local_maintained_view_subscription_from_binding_view(
-        &mut self,
-        local: &mut LocalMaintainedViewSubscription,
-        binding_view_key: BindingViewKey,
-    ) -> Result<(), Error> {
-        // Settled result sets can include support members used to maintain relations or
-        // policies. The occurrence sidecar describes only public query roots, matching
-        // the authoritative snapshot's `root_count`, so exclude those support members.
-        local.result_set = self
-            .query
-            .settled_result_sets
-            .get(&binding_view_key)
-            .map(|members| {
-                members
-                    .iter()
-                    .filter(|member| {
-                        is_public_result_member(
-                            member,
-                            local.result_table.as_str(),
-                            local.result_query.aggregate.is_some(),
-                        )
-                    })
-                    .cloned()
-                    .collect()
-            })
-            .unwrap_or_default();
-        local.authoritative_result_set = local.result_set.clone();
-        local.authoritative_result_generation =
-            self.applied_view_update_generation(binding_view_key);
-        local.program_facts = self
-            .query
-            .settled_program_facts
-            .get(&binding_view_key)
-            .cloned()
-            .unwrap_or_default();
-        if local.result_query.aggregate.is_some() {
-            local
-                .maintained
-                .replace_aggregate_result_state(&local.result_set, &local.program_facts);
-        }
-        local.result_payloads = local
-            .program_facts
-            .iter()
-            .filter_map(|fact| match fact {
-                ProgramFactEntry::ResultPayload(payload)
-                    if is_public_result_member(
-                        &payload.member,
-                        local.result_table.as_str(),
-                        local.result_query.aggregate.is_some(),
-                    ) =>
-                {
-                    Some((payload.member.clone(), payload.clone()))
-                }
-                _ => None,
-            })
-            .collect();
-        // An authoritative reset replaces membership without flowing through
-        // the ordinary local transition reducer. Rebuild the occurrence
-        // sidecar from exactly that new state before the caller pairs it with
-        // the reset snapshot; retaining the opening vector makes a later
-        // reset fail its root-count invariant (or, worse, pair wrong roots).
-        local.root_occurrence_ids = self
-            .materialize_local_maintained_relation_snapshot_with_occurrences(local)?
-            .root_occurrence_ids;
-        Ok(())
-    }
-
-    pub(crate) fn seed_local_maintained_authoritative_result_membership(
-        &self,
-        local: &mut LocalMaintainedViewSubscription,
-        binding_view_key: BindingViewKey,
-    ) {
-        local.authoritative_result_set = self
-            .query
-            .settled_result_sets
-            .get(&binding_view_key)
-            .cloned()
-            .unwrap_or_default();
-        local.authoritative_result_generation =
-            self.applied_view_update_generation(binding_view_key);
-    }
-
-    fn drain_local_maintained_view_subscription_transitions(
-        &mut self,
-        local: &mut LocalMaintainedViewSubscription,
-        authoritative_binding_view: Option<BindingViewKey>,
-    ) -> Result<Option<super::maintained_subscription_view::ResultTransitions>, Error> {
-        if local.result_query.aggregate.is_some()
-            && let Some(remote_members) =
-                self.query.settled_result_sets.get(&local.binding_view_key)
-            && let Some(remote_facts) = self
-                .query
-                .settled_program_facts
-                .get(&local.binding_view_key)
-        {
-            let visible_members = remote_members
-                .iter()
-                .filter(|member| {
-                    is_public_result_member(
-                        member,
-                        local.result_table.as_str(),
-                        local.result_query.aggregate.is_some(),
-                    )
-                })
-                .cloned()
-                .collect::<BTreeSet<_>>();
-            let visible_facts = remote_facts
-                .iter()
-                .filter(|fact| match fact {
-                    ProgramFactEntry::ResultPayload(payload) => is_public_result_member(
-                        &payload.member,
-                        local.result_table.as_str(),
-                        local.result_query.aggregate.is_some(),
-                    ),
-                    _ => false,
-                })
-                .cloned()
-                .collect::<BTreeSet<_>>();
-            if visible_members != local.result_set || visible_facts != local.program_facts {
-                let mut transitions = super::maintained_subscription_view::ResultTransitions {
-                    adds: visible_members
-                        .difference(&local.result_set)
-                        .cloned()
-                        .collect(),
-                    removes: local
-                        .result_set
-                        .difference(&visible_members)
-                        .cloned()
-                        .collect(),
-                    program_fact_adds: visible_facts
-                        .difference(&local.program_facts)
-                        .cloned()
-                        .collect(),
-                    program_fact_removes: local
-                        .program_facts
-                        .difference(&visible_facts)
-                        .cloned()
-                        .collect(),
-                    ..Default::default()
-                };
-                transitions.result_payload_adds = transitions
-                    .program_fact_adds
-                    .iter()
-                    .filter_map(|fact| match fact {
-                        ProgramFactEntry::ResultPayload(payload) => {
-                            Some((payload.member.clone(), payload.clone()))
-                        }
-                        _ => None,
-                    })
-                    .collect();
-                transitions.result_payload_removes = transitions
-                    .program_fact_removes
-                    .iter()
-                    .filter_map(|fact| match fact {
-                        ProgramFactEntry::ResultPayload(payload) => Some(payload.member.clone()),
-                        _ => None,
-                    })
-                    .collect();
-                return Ok(Some(transitions));
-            }
-        }
-        let mut states = BTreeMap::<ResultMemberEntry, (bool, bool)>::new();
-        let mut payload_states = BTreeMap::<
-            ResultMemberEntry,
-            (
-                Option<ResultMemberPayloadEntry>,
-                Option<ResultMemberPayloadEntry>,
-            ),
-        >::new();
-        let mut fact_states = BTreeMap::<ProgramFactEntry, (bool, bool)>::new();
-        let mut structured_app_row_changes = BTreeSet::new();
-        let mut terminal_operations = Vec::new();
-        let mut authoritative_membership_changed = false;
-        let mut authoritative_member_adds = BTreeSet::new();
-        if let Some(binding_view) = authoritative_binding_view {
-            let authoritative_generation = self.applied_view_update_generation(binding_view);
-            // Local optimistic changes can advance the maintained graph
-            // without any newer serving-peer membership decision. Keep them
-            // visible until an authoritative generation advances.
-            if authoritative_generation != local.authoritative_result_generation {
-                let remote_members = self
-                    .query
-                    .settled_result_sets
-                    .get(&binding_view)
-                    .cloned()
-                    .unwrap_or_default();
-                let remote_payloads = self
-                    .query
-                    .settled_program_facts
-                    .get(&binding_view)
-                    .into_iter()
-                    .flatten()
-                    .filter_map(|fact| match fact {
-                        ProgramFactEntry::ResultPayload(payload) => {
-                            Some((payload.member.clone(), payload.clone()))
-                        }
-                        _ => None,
-                    })
-                    .collect::<BTreeMap<_, _>>();
-                // The local maintained graph may intentionally be behind the
-                // authority frontier (for example, an Edge-tier window over a
-                // client-local database).  An authoritative ViewUpdate is not
-                // merely a revocation signal: it is the current membership
-                // decision for that frontier.  Import members newly admitted
-                // there so a row promoted across a TopBy boundary is delivered
-                // even though none of its locally-visible source facts changed.
-                for entry in remote_members.difference(&local.authoritative_result_set) {
-                    if !local.result_set.contains(entry) {
-                        let materializable = if remote_payloads.contains_key(entry) {
-                            true
-                        } else if let Some(row) =
-                            self.materialize_local_maintained_view_result_member(local, entry)?
-                        {
-                            let table = self.table(row.table())?;
-                            current_row_has_required_subscription_cells(
-                                &row,
-                                table,
-                                local.result_select.as_deref(),
-                            )
-                        } else {
-                            false
-                        };
-                        // Authority membership can arrive before the admitted
-                        // row's readable content bundle. Do not publish a
-                        // synthetic placeholder root: the ordinary maintained
-                        // source transition will add it once that content is
-                        // locally materializable.
-                        if !materializable {
-                            continue;
-                        }
-                        authoritative_membership_changed = true;
-                        authoritative_member_adds.insert(entry.clone());
-                        states.insert(entry.clone(), (false, true));
-                        if let Some(payload) = remote_payloads.get(entry) {
-                            payload_states.insert(
-                                entry.clone(),
-                                (
-                                    local.result_payloads.get(entry).cloned(),
-                                    Some(payload.clone()),
-                                ),
-                            );
-                        }
-                    }
-                }
-                for entry in local.authoritative_result_set.difference(&remote_members) {
-                    // Replace the exact prior member even when its output
-                    // occurrence remains visible through a new content
-                    // version. The snapshot reducer coalesces the matching
-                    // occurrence add/remove into one replacement.
-                    if local.result_set.contains(entry) {
-                        authoritative_membership_changed = true;
-                        states.insert(entry.clone(), (true, false));
-                        if local.result_payloads.contains_key(entry) {
-                            payload_states.insert(
-                                entry.clone(),
-                                (local.result_payloads.get(entry).cloned(), None),
-                            );
-                        }
-                    }
-                }
-                local.authoritative_result_set = remote_members;
-                local.authoritative_result_generation = authoritative_generation;
-            }
-        }
-        loop {
-            match local.subscription.try_recv() {
-                Ok(deltas) => {
-                    let transitions = local.maintained.apply_multisink_deltas(
-                        deltas,
-                        &local.terminal_schemas,
-                        &local.tables,
-                        &self.node_aliases,
-                    )?;
-                    structured_app_row_changes.extend(transitions.structured_app_row_changes);
-                    terminal_operations.extend(transitions.terminal_operations);
-                    for entry in transitions.adds {
-                        let before = local.result_set.contains(&entry);
-                        states
-                            .entry(entry)
-                            .and_modify(|(_, after)| *after = true)
-                            .or_insert((before, true));
-                    }
-                    for entry in transitions.removes {
-                        let before = local.result_set.contains(&entry);
-                        states
-                            .entry(entry)
-                            .and_modify(|(_, after)| *after = false)
-                            .or_insert((before, false));
-                    }
-                    for member in transitions.result_payload_removes {
-                        let before = local.result_payloads.get(&member).cloned();
-                        payload_states
-                            .entry(member)
-                            .and_modify(|(_, after)| *after = None)
-                            .or_insert((before, None));
-                    }
-                    for (member, payload) in transitions.result_payload_adds {
-                        let before = local.result_payloads.get(&member).cloned();
-                        payload_states
-                            .entry(member)
-                            .and_modify(|(_, after)| *after = Some(payload.clone()))
-                            .or_insert((before, Some(payload)));
-                    }
-                    for fact in transitions.program_fact_adds {
-                        let before = local.program_facts.contains(&fact);
-                        fact_states
-                            .entry(fact)
-                            .and_modify(|(_, after)| *after = true)
-                            .or_insert((before, true));
-                    }
-                    for fact in transitions.program_fact_removes {
-                        let before = local.program_facts.contains(&fact);
-                        fact_states
-                            .entry(fact)
-                            .and_modify(|(_, after)| *after = false)
-                            .or_insert((before, false));
-                    }
-                }
-                Err(std::sync::mpsc::TryRecvError::Empty) => break,
-                Err(std::sync::mpsc::TryRecvError::Disconnected) => {
-                    return Err(Error::SubscriptionClosed);
-                }
-            }
-        }
-        if states.is_empty()
-            && payload_states.is_empty()
-            && fact_states.is_empty()
-            && structured_app_row_changes.is_empty()
-            && terminal_operations.is_empty()
-        {
-            return Ok(None);
-        }
-        let mut transitions = super::maintained_subscription_view::ResultTransitions {
-            authoritative_membership_changed,
-            authoritative_member_adds,
-            structured_app_row_changes,
-            terminal_operations,
-            ..Default::default()
-        };
-        for (entry, (before, after)) in states {
-            match (before, after) {
-                (false, true) => transitions.adds.push(entry),
-                (true, false) => transitions.removes.push(entry),
-                _ => {}
-            }
-        }
-        for (member, (before, after)) in payload_states {
-            match (before, after) {
-                (None, Some(payload)) => transitions.result_payload_adds.push((member, payload)),
-                (Some(_), None) => transitions.result_payload_removes.push(member),
-                (Some(before), Some(after)) if before != after => {
-                    transitions.result_payload_removes.push(member.clone());
-                    transitions.result_payload_adds.push((member, after));
-                }
-                _ => {}
-            }
-        }
-        for (fact, (before, after)) in fact_states {
-            match (before, after) {
-                (false, true) => transitions.program_fact_adds.push(fact),
-                (true, false) => transitions.program_fact_removes.push(fact),
-                _ => {}
-            }
-        }
-        // Aggregate facts are the payload vocabulary for synthetic result
-        // members. Preserve their current values alongside membership while
-        // coalescing a multisink batch; otherwise a present NULL or a revised
-        // aggregate can be mistaken for an absent payload at materialization.
-        if local.result_query.aggregate.is_some() {
-            transitions.result_payload_adds = transitions
-                .program_fact_adds
-                .iter()
-                .filter_map(|fact| match fact {
-                    ProgramFactEntry::ResultPayload(payload) => {
-                        Some((payload.member.clone(), payload.clone()))
-                    }
-                    _ => None,
-                })
-                .collect();
-            transitions.result_payload_removes = transitions
-                .program_fact_removes
-                .iter()
-                .filter_map(|fact| match fact {
-                    ProgramFactEntry::ResultPayload(payload) => Some(payload.member.clone()),
-                    _ => None,
-                })
-                .collect();
-        }
-        Ok(Some(transitions))
-    }
-
-    fn apply_local_maintained_view_transitions(
-        &mut self,
-        local: &mut LocalMaintainedViewSubscription,
-        transitions: super::maintained_subscription_view::ResultTransitions,
-    ) -> Result<LocalMaintainedViewSubscriptionUpdate, Error> {
-        self.apply_local_maintained_view_transitions_inner(local, transitions, true)
-    }
-
-    fn apply_local_maintained_view_transitions_inner(
-        &mut self,
-        local: &mut LocalMaintainedViewSubscription,
-        transitions: super::maintained_subscription_view::ResultTransitions,
-        materialize_update: bool,
-    ) -> Result<LocalMaintainedViewSubscriptionUpdate, Error> {
-        let structured_output = !local.result_query.array_subqueries.is_empty();
-        let authoritative_membership_changed = transitions.authoritative_membership_changed;
-        let authoritative_member_adds = transitions.authoritative_member_adds;
-        let structured_app_row_changes = transitions.structured_app_row_changes.clone();
-        let terminal_operations = transitions.terminal_operations.clone();
-        let terminal_layout = (!terminal_operations.is_empty())
-            .then(|| local.terminal_schemas.terminal_root_layout().cloned())
-            .flatten();
-        let aggregate_replacements = transitions
-            .adds
-            .iter()
-            .filter(|member| {
-                is_public_aggregate_result_member(
-                    member,
-                    local.result_table.as_str(),
-                    local.result_query.aggregate.is_some(),
-                )
-            })
-            .map(aggregate_result_member_row_uuid)
-            .collect::<Result<BTreeSet<_>, _>>()?;
-        let mut added = Vec::new();
-        let mut removed = Vec::new();
-        let mut added_edges = Vec::new();
-        let mut removed_edges = Vec::new();
-        for member in transitions.result_payload_removes {
-            local.result_payloads.remove(&member);
-        }
-        for (member, payload) in transitions.result_payload_adds {
-            if is_public_result_member(
-                &member,
-                local.result_table.as_str(),
-                local.result_query.aggregate.is_some(),
-            ) {
-                local.result_payloads.insert(member, payload);
-            }
-        }
-        for member in transitions.adds {
-            if !is_public_result_member(
-                &member,
-                local.result_table.as_str(),
-                local.result_query.aggregate.is_some(),
-            ) {
-                continue;
-            }
-            // Authority-scope re-entry can surface a new internal result
-            // member for an occurrence the public facade already tracks. In
-            // that case replace the stale member so the current scope wins.
-            // Ordinary content updates keep their member identity: replacing
-            // them here turns an update into a duplicate add and leaves the
-            // public subscription with the old payload.
-            replace_stale_authoritative_occurrence_member(
-                &mut local.result_set,
-                &mut local.result_payloads,
-                &authoritative_member_adds,
-                &member,
-                local.result_table.as_str(),
-                local.result_query.aggregate.is_some(),
-            )?;
-            if local.result_set.insert(member.clone()) && materialize_update && !structured_output {
-                if let Some(row) =
-                    self.materialize_local_maintained_view_result_member(local, &member)?
-                    && let Some(occurrence_id) = public_result_member_occurrence_id(
-                        &member,
-                        local.result_table.as_str(),
-                        local.result_query.aggregate.is_some(),
-                    )?
-                {
-                    added.push((occurrence_id, row));
-                }
-            }
-        }
-        for member in transitions.removes {
-            if !is_public_result_member(
-                &member,
-                local.result_table.as_str(),
-                local.result_query.aggregate.is_some(),
-            ) {
-                continue;
-            }
-            if local.result_set.remove(&member) {
-                if materialize_update && !structured_output {
-                    if let Some(occurrence_id) = member.output_occurrence_id() {
-                        removed.push(occurrence_id);
-                    } else if is_public_aggregate_result_member(
-                        &member,
-                        local.result_table.as_str(),
-                        local.result_query.aggregate.is_some(),
-                    ) {
-                        let row_uuid = aggregate_result_member_row_uuid(&member)?;
-                        let replacement_is_current = local.result_set.iter().any(|candidate| {
-                            is_public_aggregate_result_member(
-                                candidate,
-                                local.result_table.as_str(),
-                                local.result_query.aggregate.is_some(),
-                            ) && aggregate_result_member_row_uuid(candidate)
-                                .is_ok_and(|candidate_uuid| candidate_uuid == row_uuid)
-                        });
-                        if !aggregate_replacements.contains(&row_uuid) && !replacement_is_current {
-                            removed.push(OutputOccurrenceId::single_source(ObjectId::from_uuid(
-                                row_uuid.0,
-                            )));
-                        }
-                    }
-                }
-            }
-        }
-        for fact in transitions.program_fact_removes {
-            if local.program_facts.remove(&fact) {
-                if materialize_update
-                    && !structured_output
-                    && let ProgramFactEntry::RelationEdge(edge) = fact
-                {
-                    removed_edges.push(self.project_relation_edge_through_read_schema(
-                        &edge,
-                        local.result_schema_version,
-                    )?);
-                }
-            }
-        }
-        for fact in transitions.program_fact_adds {
-            let edge = (materialize_update && !structured_output)
-                .then(|| match &fact {
-                    ProgramFactEntry::RelationEdge(edge) => Some(edge.clone()),
-                    _ => None,
-                })
-                .flatten();
-            if local.program_facts.insert(fact)
-                && let Some(edge) = edge
-            {
-                let relation_edge = self.project_relation_edge_through_read_schema(
-                    &edge,
-                    local.result_schema_version,
-                )?;
-                let row = if let Some(version) = &edge.target_version {
-                    self.materialize_local_maintained_view_relation_edge_row(
-                        local,
-                        edge.target_table.as_str(),
-                        edge.target_row,
-                        version.tx,
-                    )?
-                } else {
-                    None
-                };
-                added_edges.push((relation_edge, row));
-            }
-        }
-        if materialize_update && structured_output {
-            for root in structured_app_row_changes {
-                match local.maintained.structured_app_row(root) {
-                    Some(record) => added.push((
-                        OutputOccurrenceId::single_source(ObjectId::from_uuid(root.0)),
-                        CurrentRow::new(local.result_table.clone(), record),
-                    )),
-                    None => removed.push(OutputOccurrenceId::single_source(ObjectId::from_uuid(
-                        root.0,
-                    ))),
-                }
-            }
-        }
-        Ok(LocalMaintainedViewSubscriptionUpdate {
-            authoritative_membership_changed,
-            added,
-            removed,
-            added_edges,
-            removed_edges,
-            terminal_operations,
-            terminal_layout,
-        })
     }
 
     fn bounded_historical_current_rows(
